Added authorization on updatee and delete

+++ app/Http/Controllers/QuestionController.php
@@ -6,6 +6,7 @@ use App\Http\Requests\AskQuestionRequest;
 use App\Models\Question;
 use Illuminate\Foundation\Http\FormRequest;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class QuestionController extends Controller
 {
@@ -63,6 +64,9 @@ class QuestionController extends Controller
      */
     public function edit(Question $question)
     {
+        if (Gate::denies('update-question', $question)) {
+            abort(403, "Access Denied");
+        }
         return view('questions.edit', compact('question'));
     }
 
@@ -75,6 +79,9 @@ class QuestionController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if (Gate::denies('update-question', $question)) {
+            abort(403, "Access Denied");
+        }
         $question->update($request->only('title', 'body'));
         return redirect('/questions')->with('success', "Your Question Successfully Updated");
     }
@@ -87,6 +94,9 @@ class QuestionController extends Controller
      */
     public function destroy(Question $question)
     {
+        if (Gate::denies('delete-question', $question)) {
+            abort(403, "Access Denied");
+        }
         $question->delete();
         return redirect('/questions')->with('success', "Quesion deleted successfully");
     }
